Create Shiny1_Vasudevamurthy.Rmd
This shiny app is developed to consolidate data from excel files that contain information on baby names for both boys and girls from the year 1954 to 2018.
This application contains 4 individual tabs, each providing its own interactive functionality. The first two tabs provide the option to pull out the top 10 baby names for any year between 1954 to 2018 for boys and girls respectively. Each tab is complimented by a drop-down menu in the sidebar panel where the user can select the year of choice to view the top 10 baby names.
The next two tabs provide the user the ability to view trends for a specific name from 1954 to 2018 for boys and girls respectively. Again, here each tab is complemented by a drop-down menu in the sidebar panel where the user can select their name of choice to view the popularity trend between 1954 to 2018 for that name.
The first part of the code includes cleaning the data, reformatting and transposing it to bring the data to the format that is conducive to the coding of the shiny application.
The second half of the code is for the shiny application itself and begins with the coding of the ui interface where the layout of the application is defined to include the sidebar and the mainpanel constructs. This is followed by the server function where all the outputs are defined and linked to the tabs to obtain the input provided by the user.
Finally, the shinyApp function brings together the ui and server elements to launch the shiny application.